made table adjustments and created filter modal component
